[#67] Extract functionality to extension methods
