+++ app/src/main/java/net/aiscope/gdd_app/ui/metadata/MetadataActivity.kt
@@ -3,18 +3,10 @@ package net.aiscope.gdd_app.ui.metadata
 import android.content.Intent
 import android.graphics.Typeface
 import android.os.Bundle
-import android.text.SpannableStringBuilder
-import android.text.style.ImageSpan
 import android.widget.AbsSpinner
-import android.widget.TextView
 import android.widget.Toast
-import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.text.color
-import androidx.core.text.scale
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.google.android.material.snackbar.BaseTransientBottomBar.BaseCallback
 import com.google.android.material.snackbar.Snackbar
 import dagger.android.AndroidInjection
 import kotlinx.android.synthetic.main.activity_metadata.*
@@ -24,6 +16,9 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
 import net.aiscope.gdd_app.R
+import net.aiscope.gdd_app.extensions.buildSnackbarContent
+import net.aiscope.gdd_app.extensions.setActionFontType
+import net.aiscope.gdd_app.extensions.setActionOnDismissed
 import net.aiscope.gdd_app.ui.CaptureFlow
 import net.aiscope.gdd_app.ui.attachCaptureFlowToolbar
 import net.aiscope.gdd_app.ui.capture.CaptureImageActivity
@@ -103,17 +98,13 @@ class MetadataActivity : AppCompatActivity() , MetadataView, CaptureFlow {
     }
 
     private fun showRetryBar() {
-        val snackbar = Snackbar.make(
+        Snackbar.make(
             findViewById(android.R.id.content),
             buildSnackbarContent(R.string.metadata_snackbar_error),
             Snackbar.LENGTH_LONG)
             .setAction(R.string.metadata_snackbar_retry) { save() }
-
-        val snackbarActionTextView =
-            snackbar.view.findViewById(R.id.snackbar_action) as TextView
-        snackbarActionTextView.setTypeface(snackbarActionTextView.typeface, Typeface.BOLD)
-
-        snackbar.show()
+            .setActionFontType(Typeface.BOLD)
+            .show()
     }
 
     private fun finishFlow() {
@@ -121,27 +112,9 @@ class MetadataActivity : AppCompatActivity() , MetadataView, CaptureFlow {
             findViewById(android.R.id.content),
             buildSnackbarContent(R.string.metadata_snackbar_success,R.drawable.ic_thumb_up),
             Snackbar.LENGTH_SHORT)
-            .addCallback(object :
-                BaseCallback<Snackbar?>() {
-                override fun onDismissed(transientBottomBar: Snackbar?, event: Int) {
-                    goToHome()
-                    super .onDismissed(transientBottomBar, event)
-                }
-            })
+            .setActionOnDismissed { goToHome() }
             .show()
     }
-
-    private fun buildSnackbarContent(@StringRes text: Int, @DrawableRes icon: Int? = null): SpannableStringBuilder {
-        val builder = SpannableStringBuilder()
-        if(icon != null) {
-            builder.append("  ")
-            builder.setSpan(ImageSpan(this, icon), builder.length - 1, builder.length, 0)
-            builder.append("  ")
-        }
-        builder.color(getColor(R.color.colorPlainText)) { scale(1.2f) { append(getString(text)) } }
-        return builder
-    }
-
 
     override fun captureImage(nextImageName: String, nextMaskName: String) {
         val intent = Intent(this, CaptureImageActivity::class.java)
